backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0

# Vector Store & Embeddings
chromadb>=0.4.0
openai>=1.0.0
tiktoken>=0.5.0

# LLM & Streaming
anthropic>=0.18.0
sse-starlette>=1.6.0

# Async file operations
aiofiles>=23.0.0

# Settings management
pydantic-settings>=2.0.0
